refactor: create a python package
