--- requirements.txt
@@ -1,3 +1,0 @@
-flask
-flask_cors
-requests
