+++ src/gitlab_pipeline_visualizer/web/__main__.py
@@ -1,0 +1,4 @@
+from . import cli
+
+if __name__ == "__main__":
+    cli()
