+++ src/gitlab_pipeline_visualizer/web/templates/index.html
